v0.1.10 - Los siguientes métodos se traspasan de Util a AppService: 	- String getAppNameByPackageName(String packageName). 	- Drawable getDrawableByPackageName(String packageName). 	- boolean isSystemPackage(PackageInfo pkgInfo).
