+++ app/src/main/java/org/prezdev/notihistory/service/NotificationService.java
@@ -11,5 +11,5 @@ public interface NotificationService {
 
     List<NotificationVO> findAllByPackageName(String packageName);
 
-    List<InstalledApp> getInstalledApps(boolean systemAppsIncluded);
+
 }
