+++ app/src/main/java/org/prezdev/notihistory/service/impl/NotificationServiceImpl.java
@@ -44,57 +44,4 @@ public class NotificationServiceImpl implements NotificationService {
     public List<NotificationVO> findAllByPackageName(String packageName) {
         return notificationDao.findAllByPackageName(packageName);
     }
-
-    @Override
-    public List<InstalledApp> getInstalledApps(boolean systemAppsIncluded) {
-        List<InstalledApp> installedApps = new ArrayList<>();
-        PackageManager packageManager = this.context.getPackageManager();
-
-        InstalledApp installedApp;
-        String appName;
-        String packageName;
-        Util util = Util.getInstance(packageManager);
-
-        List<PackageInfo> packs = packageManager.getInstalledPackages(0);
-
-        int id = 1;
-        for(PackageInfo packageInfo : packs){
-            // No hay que incluir las apps del systema?
-            if(!systemAppsIncluded){
-                // la app es del sistema
-                if(util.isSystemPackage(packageInfo)){
-                    continue;
-                }
-            }
-
-            installedApp = new InstalledApp();
-
-            packageName = packageInfo.packageName;
-
-            installedApp.setId(id++);
-            installedApp.setPackageName(packageName);
-            appName = util.getAppNameByPackageName(packageName);
-            installedApp.setName(appName);
-
-            try {
-                installedApp.setIcon(util.getDrawableByPackageName(packageName));
-            } catch (PackageManager.NameNotFoundException e) { }
-
-            installedApp.setVersionName(packageInfo.versionName);
-            installedApp.setVersionCode(packageInfo.versionCode);
-
-            installedApps.add(installedApp);
-        }
-
-        Comparator<InstalledApp> installedAppComparator = new Comparator<InstalledApp>() {
-            @Override
-            public int compare(InstalledApp installedApp1, InstalledApp installedApp2) {
-                return installedApp1.getName().compareTo(installedApp2.getName());
-            }
-        };
-
-        Collections.sort(installedApps, installedAppComparator);
-
-        return installedApps;
-    }
 }
